Update error.html
The error.html file is a custom HTML page that is displayed when a user encounters an error while accessing website. In the context of hosting a static website on AWS S3, this file is often used to handle common HTTP error responses like 404 (Not Found) or 500 (Internal Server Error).