Update index.html
The index.html file is the default landing page for website. It is the first page that users see when they visit the website's root URL (e.g., http://example.com/). In a static website hosted on platforms like AWS S3, the index.html file serves as the main entry point of the site.